Remove redis dependency for OGN stats
